修改消息接口代码

+++ find-common/src/main/java/com/potato369/find/common/vo/LikesMessageVO.java
@@ -13,14 +13,14 @@ import lombok.*;
 @EqualsAndHashCode(callSuper = false)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class LikesMessageVO {
-	
+
     //最新一条未读点赞消息内容
     @JSONField(name = "content1")
     @JsonProperty(value = "content1")
     private String content;
-    
+
     //未读点赞消息总条数
     @JSONField(name = "count1")
     @JsonProperty(value = "count1")
-    private Integer count;
+    private Long count;
 }
